Added prints

+++ prepare-market.py
@@ -16,7 +16,7 @@ save_path = download_path + '/pytorch'
 if not os.path.isdir(save_path):
     os.mkdir(save_path)
 #-----------------------------------------
-#query
+print("query")
 query_path = download_path + '/query'
 query_save_path = download_path + '/pytorch/query'
 if not os.path.isdir(query_save_path):
@@ -24,17 +24,20 @@ if not os.path.isdir(query_save_path):
 
 for root, dirs, files in os.walk(query_path, topdown=True):
     for name in files:
+
         if not name[-3:]=='jpg':
             continue
         ID  = name.split('_')
         src_path = query_path + '/' + name
-        dst_path = query_save_path + '/' + ID[0] 
-        if not os.path.isdir(dst_path):
-            os.mkdir(dst_path)
-        copyfile(src_path, dst_path + '/' + name)
+        dst_path = query_save_path + '/' + ID[0]
+
+        if not os.path.exists(dst_path + '/' + name):
+            if not os.path.isdir(dst_path):
+                os.mkdir(dst_path)
+            copyfile(src_path, dst_path + '/' + name)
 
 #-----------------------------------------
-#multi-query
+print("multi")
 query_path = download_path + '/gt_bbox'
 # for dukemtmc-reid, we do not need multi-query
 if os.path.isdir(query_path):
@@ -49,12 +52,13 @@ if os.path.isdir(query_path):
             ID  = name.split('_')
             src_path = query_path + '/' + name
             dst_path = query_save_path + '/' + ID[0]
-            if not os.path.isdir(dst_path):
-                os.mkdir(dst_path)
-            copyfile(src_path, dst_path + '/' + name)
+            if not os.path.exists(dst_path + '/' + name):
+                if not os.path.isdir(dst_path):
+                    os.mkdir(dst_path)
+                copyfile(src_path, dst_path + '/' + name)
 
 #-----------------------------------------
-#gallery
+print("#gallery")
 gallery_path = download_path + '/bounding_box_test'
 gallery_save_path = download_path + '/pytorch/gallery'
 if not os.path.isdir(gallery_save_path):
@@ -67,12 +71,13 @@ for root, dirs, files in os.walk(gallery_path, topdown=True):
         ID  = name.split('_')
         src_path = gallery_path + '/' + name
         dst_path = gallery_save_path + '/' + ID[0]
-        if not os.path.isdir(dst_path):
-            os.mkdir(dst_path)
-        copyfile(src_path, dst_path + '/' + name)
+        if not os.path.exists(dst_path + '/' + name):
+            if not os.path.isdir(dst_path):
+                os.mkdir(dst_path)
+            copyfile(src_path, dst_path + '/' + name)
 
 #---------------------------------------
-#train_all
+print("#train_all")
 train_path = download_path + '/bounding_box_train'
 train_save_path = download_path + '/pytorch/train_all'
 if not os.path.isdir(train_save_path):
@@ -85,13 +90,14 @@ for root, dirs, files in os.walk(train_path, topdown=True):
         ID  = name.split('_')
         src_path = train_path + '/' + name
         dst_path = train_save_path + '/' + ID[0]
-        if not os.path.isdir(dst_path):
-            os.mkdir(dst_path)
-        copyfile(src_path, dst_path + '/' + name)
+        if not os.path.exists(dst_path + '/' + name):
+            if not os.path.isdir(dst_path):
+                os.mkdir(dst_path)
+            copyfile(src_path, dst_path + '/' + name)
 
 
 #---------------------------------------
-#train_val
+print("#train_all")
 train_path = download_path + '/bounding_box_train'
 train_save_path = download_path + '/pytorch/train'
 val_save_path = download_path + '/pytorch/val'
@@ -106,8 +112,9 @@ for root, dirs, files in os.walk(train_path, topdown=True):
         ID  = name.split('_')
         src_path = train_path + '/' + name
         dst_path = train_save_path + '/' + ID[0]
-        if not os.path.isdir(dst_path):
-            os.mkdir(dst_path)
-            dst_path = val_save_path + '/' + ID[0]  #first image is used as val image
-            os.mkdir(dst_path)
-        copyfile(src_path, dst_path + '/' + name)
+        if not os.path.exists(dst_path + '/' + name):
+            if not os.path.isdir(dst_path):
+                os.mkdir(dst_path)
+                dst_path = val_save_path + '/' + ID[0]  #first image is used as val image
+                os.mkdir(dst_path)
+            copyfile(src_path, dst_path + '/' + name)
